se esta acomodando lo de las validaciones

+++ src/main/java/com/cristofer/apirest/apirest/Entities/DetalleVenta.java
@@ -30,7 +30,7 @@ public class DetalleVenta {
 
     private int cantidad;
 
-    private double subtotal;
+    private int subtotal;
 
-    private double total;
+    private int total;
 }
